ml_model/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
autograd==1.6.2
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.6.3
blis==0.7.11
bs4==0.0.1
cached-property==1.5.2
cachetools==5.3.1
catalogue==2.0.10
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.15.1
comm==0.1.3
commonmark==0.9.1
confection==0.1.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dictdiffer==0.9.0
executing==1.2.0
fastjsonschema==2.18.0
Flask
flatbuffers==23.5.26
fonttools==4.41.0
fqdn==1.5.1
future==0.18.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.9.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.14.0
kiwisolver==1.4.4
langcodes==3.3.0
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==3.0.1
ml-dtypes==0.2.0
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==7.0.4
notebook_shim==0.2.3
notion==0.0.28
numpy==1.25.1
oauthlib==3.2.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickle-mixin==1.0.2
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
pluggy==1.3.0
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.38
protobuf==4.24.4
psutil==5.9.5
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydotplus==2.0.2
Pygments==2.15.1
pyparsing==3.0.9
pytest==7.4.2
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==7.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.1
Send2Trash==1.8.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.2
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.3.0
urllib3==1.26.13
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug
widgetsnbextension==4.0.9
wrapt==1.14.1
